requirements.txt
fastapi
uvicorn
jinja2
pandas
numpy
python-dotenv
requests
beautifulsoup4
